+++ writting-tests/with-tap/routes/pets.js
@@ -1,0 +1,5 @@
+export default async function (app, opts) {
+  app.get('/', async () => {
+    return { hello: 'world' }
+  })
+}
